foundation/rag/requirements.txt
fastapi
uvicorn[standard]
haystack-ai
qdrant-haystack
python-dotenv
